Alot of changes in this commit:
General:
    - Added override keyword on some overrided methods
    - Added Drawable class to allow custom draw any class
    - Added packed versions of Vector2, Vector3 and Vector4
    - Added UniformBufferObject
    - Changed Vertex to a struct
    - Added LoadModel test

OpenGL:
    - Set OpenGL min version to 4.2
    - SetUniformBufferObject added as part of GL_Shader

Vulkan:
    - Added initial implementation of Vk_Mesh

+++ src/Graphics/ImageLoader.cpp
@@ -20,7 +20,8 @@ bool ImageLoader::LoadFromFile(const String& filename,
     std::vector<byte> out;
     FileSystem& fs = FileSystem::GetInstance();
     if (fs.LoadFileData(filename, &out)) {
-        return LoadFromFileInMemory(out.data(), out.size(), pixels, size);
+        return LoadFromFileInMemory(out.data(), static_cast<uint32>(out.size()),
+                                    pixels, size);
     }
     return false;
 }
